Move comment display rendering to a service.
