+++ Models/MappingProfiles/CommentMappingProfile.cs
@@ -21,23 +21,32 @@ along with Palaver.  If not, see <http://www.gnu.org/licenses/>.
 using System;
 using AutoMapper;
 using PalaverCore.Models.CommentViewModels;
+using PalaverCore.Services;
 
 namespace PalaverCore.Models.MappingProfiles
 {
     public class CommentMappingProfile : Profile
     {
-        public CommentMappingProfile()
+
+        private CommentRenderService _commentRenderer;
+
+        public CommentMappingProfile(CommentRenderService commentRenderer)
         {
+            this._commentRenderer = commentRenderer;
+
             CreateMap<Comment, CreateViewModel>();
             CreateMap<Comment, DetailViewModel>()
                 .ForMember(d => d.UserName, opt => opt.MapFrom(s => s.User.UserName))
                 .ForMember(d => d.EmailHash, opt => opt.MapFrom(s => s.User.EmailHash))
-                .ForMember(d => d.Url, opt => opt.MapFrom(s => $"{Startup.SiteRoot}/Thread/{s.ThreadId}/{s.Id}"));
+                .ForMember(d => d.Url, opt => opt.MapFrom(s => $"{Startup.SiteRoot}/Thread/{s.ThreadId}/{s.Id}"))
+                .ForMember(d => d.DisplayText, opt => opt.MapFrom(s => _commentRenderer.ToHtml(s.Text, s.Format)));
             CreateMap<Comment, EditViewModel>();
-            CreateMap<Comment, EditResultViewModel>();
+            CreateMap<Comment, EditResultViewModel>()
+                .ForMember(d => d.DisplayText, opt => opt.MapFrom(s => _commentRenderer.ToHtml(s.Text, s.Format)));
             CreateMap<Comment, SearchResultViewModel>()
                 .ForMember(d => d.Title, opt => opt.MapFrom(s => $"[{s.CreatedDisplay}] {s.User.UserName} - {s.Thread.Title}"))
-                .ForMember(d => d.Url, opt => opt.MapFrom(s => $"{Startup.SiteRoot}/Thread/{s.ThreadId}/{s.Id}"));
+                .ForMember(d => d.Url, opt => opt.MapFrom(s => $"{Startup.SiteRoot}/Thread/{s.ThreadId}/{s.Id}"))
+                .ForMember(d => d.DisplayText, opt => opt.MapFrom(s => _commentRenderer.ToHtml(s.Text, s.Format)));
         }
     }
 }
